looks like it mostly works.

+++ source/FlaxIR/ConstantValue.cpp
@@ -154,7 +154,7 @@ namespace fir
 
 	ConstantFP::ConstantFP(Type* type, float val) : fir::ConstantValue(type)
 	{
-		this->value = val;
+		this->value = (double) val;
 	}
 
 	ConstantFP::ConstantFP(Type* type, double val) : fir::ConstantValue(type)
